+++ examples/HTTP/client.php
@@ -3,10 +3,13 @@
 include __DIR__ . '../../../vendor/autoload.php';
 
 use Rubix\Server\RESTClient;
+use Rubix\Server\HTTP\Middleware\Client\BasicAuthenticator;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 
-$client = new RESTClient('127.0.0.1', 8000);
+$client = new RESTClient('127.0.0.1', 8000, false, [
+    new BasicAuthenticator('user', 'password'),
+]);
 
 $generator = new Agglomerate([
     'red' => new Blob([255, 0, 0], 20.0),