+++ examples/HTTP/server.php
@@ -6,7 +6,8 @@ use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 use Rubix\ML\Classifiers\KDNeighbors;
 use Rubix\Server\HTTPServer;
-use Rubix\Server\HTTP\Middleware\AccessLogGenerator;
+use Rubix\Server\HTTP\Middleware\Server\AccessLogGenerator;
+use Rubix\Server\HTTP\Middleware\Server\BasicAuthenticator;
 use Rubix\ML\Other\Loggers\Screen;
 
 $generator = new Agglomerate([
@@ -23,6 +24,9 @@ $estimator->train($dataset);
 
 $server = new HTTPServer('127.0.0.1', 8000, null, [
     new AccessLogGenerator(new Screen()),
+    new BasicAuthenticator([
+        'user' => 'password',
+    ]),
 ]);
 
 $server->setLogger(new Screen());