+++ src/HTTP/Middleware/Client/Middleware.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rubix\Server\HTTP\Middleware\Client;
+
+interface Middleware
+{
+    /**
+     * Return the higher-order function.
+     *
+     * @return callable
+     */
+    public function __invoke() : callable;
+}