+++ tests/HTTP/Middleware/Server/SharedTokenAuthenticatorTest.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Rubix\Server\Tests\HTTP\Middleware;
+namespace Rubix\Server\Tests\HTTP\Middleware\Server;
 
-use Rubix\Server\HTTP\Middleware\Middleware;
-use Rubix\Server\HTTP\Middleware\SharedTokenAuthenticator;
+use Rubix\Server\HTTP\Middleware\Server\Middleware;
+use Rubix\Server\HTTP\Middleware\Server\SharedTokenAuthenticator;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @group Middleware
- * @covers \Rubix\Server\HTTP\Middleware\SharedTokenAuthenticator
+ * @covers \Rubix\Server\HTTP\Middleware\Server\SharedTokenAuthenticator
  */
 class SharedTokenAuthenticatorTest extends TestCase
 {
     /**
-     * @var \Rubix\Server\HTTP\Middleware\SharedTokenAuthenticator
+     * @var \Rubix\Server\HTTP\Middleware\Server\SharedTokenAuthenticator
      */
     protected $middleware;