+++ tests/HTTP/Middleware/Server/TrustedClientsTest.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Rubix\Server\Tests\HTTP\Middleware;
+namespace Rubix\Server\Tests\HTTP\Middleware\Server;
 
-use Rubix\Server\HTTP\Middleware\Middleware;
-use Rubix\Server\HTTP\Middleware\TrustedClients;
+use Rubix\Server\HTTP\Middleware\Server\Middleware;
+use Rubix\Server\HTTP\Middleware\Server\TrustedClients;
 use PHPUnit\Framework\TestCase;
 use Rubix\Server\Exceptions\InvalidArgumentException;
 
 /**
  * @group Middleware
- * @covers \Rubix\Server\HTTP\Middleware\TrustedClients
+ * @covers \Rubix\Server\HTTP\Middleware\Server\TrustedClients
  */
 class TrustedClientsTest extends TestCase
 {
     /**
-     * @var \Rubix\Server\HTTP\Middleware\TrustedClients
+     * @var \Rubix\Server\HTTP\Middleware\Server\TrustedClients
      */
     protected $middleware;